transform margin discussions into components
