+++ app/directives/index.ts
@@ -6,8 +6,6 @@ import elastic from './elastic';
 import elementPosition from './elementPosition';
 import elementSize from './elementSize';
 import highlights from './highlights';
-import marginDiscussion from './marginDiscussion';
-import marginDiscussionDraft from './marginDiscussionDraft';
 import inlineEditable from './inline-editable';
 import kramjax from './kramjax';
 import onOutside from './onOutside';
@@ -28,8 +26,6 @@ export default function(app) {
   elementPosition(app);
   elementSize(app);
   highlights(app);
-  marginDiscussion(app);
-  marginDiscussionDraft(app);
   inlineEditable(app);
   kramjax(app);
   onOutside(app);
